requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dateutil==2.8.2
gunicorn==21.2.0
